+++ tests/data/test_wc_1.txt
@@ -1,0 +1,5 @@
+some text
+cat cat cat cat
+dog dog dog
+
+or cat?
