+++ tests/data/test_wc_2.txt
@@ -1,0 +1,2 @@
+Somebody once told me the world is gonna roll me
+I ain't the sharpest tool in the shed
