--- tests/test2.txt
@@ -1,2 +1,0 @@
-Somebody once told me the world is gonna roll me
-I ain't the sharpest tool in the shed
